refactor: change to a pre-stored string approach
change to a pre-stored string approach because of memory usage.

BREAKING CHANGE: pre-stored string approach
Before: car.status()
After: only outputView.showStatus(), remove the car.status()
