+++ src/main/kotlin/racingcar/OutputView.kt
@@ -1,21 +1,11 @@
 package racingcar
 
-import camp.nextstep.edu.missionutils.Console
-
-object RacingView {
-    fun getNamesFromUser(): List<String> {
-        println("경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)")
-        return Console.readLine().split(",")
-    }
-
-    fun getAttemptFromUser(): Int {
-        println("시도할 횟수는 몇 회인가요?")
-        return Console.readLine().toInt()
-    }
+data class OutputView(val attempt: Attempt) {
+    val path = (0..attempt.value).map { "-".repeat(it) }
 
     private fun showStatus(round: Round) {
         for (car in round) {
-            println(car.status())
+            println(car.name + " : " + path[car.position])
         }
     }
