+++ src/test/kotlin/racingcar/CarTest.kt
@@ -46,13 +46,4 @@ class CarTest {
             assertThrows<IllegalArgumentException> { Car("pobi", -1) }
         }
     }
-
-    @ParameterizedTest
-    @CsvSource("pobi, 0, ''", "jun, 1, -", "woni, 2, --")
-    fun `자동차의 현재 상태 출력이 올바른 형식을 따르는 지 확인한다`(name: String, position: Int, output: String) {
-        assertSimpleTest {
-            val car = Car(name, position)
-            assertThat(car.status()).isEqualTo("$name : $output")
-        }
-    }
 }
